backend/requirements.txt
aiofiles
alembic
anyio
asgiref
async-generator
attrs
autopep8
bcrypt
cffi
charset-normalizer
chart-studio
click
colorama
confection
cryptography
cymem
ecdsa
exceptiongroup
fastapi
future
h11
idna
interchange
itsdangerous
Jinja2
joblib
langcodes
lxml
MarkupSafe
methodtools
monotonic
murmurhash
neo4j
nltk
numpy
outcome
packaging
pandas
pansi
pathy
Pillow
preshed
py2neo
pyasn1
pycodestyle
pycparser
pydantic
Pygments
pyinstrument
pyparsing
PySocks
python-dateutil
python-dotenv
python-jose
python-multipart
python-pptx
pytz
regex
requests
retrying
rsa
scikit-learn
scipy
sd-material-ui
selenium
six
smart-open
snakeviz
sniffio
sortedcontainers
SQLAlchemy
starlette
tenacity
toml
tornado
tqdm
trio
trio-websocket
typer
typing-extensions
ua-parser
urllib3
uvicorn
wasabi
Werkzeug
wirerope
wsproto
XlsxWriter
neo4j
py2neo
uvicorn
spacy